server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
boto3==1.29.3
openai-whisper==20231117
torch==2.1.1
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.1 